refactor: centralize http fetch logic in desktop client
