--- src/main/internal/fetcher.ts
@@ -1,117 +1,0 @@
-import { appVersion } from "@main/const";
-import { desktop } from "@main/index";
-import ky from "ky";
-import { Agent, Pool, ProxyAgent } from "undici";
-
-let cachedAgent: ProxyAgent | null = null;
-let cachedProxyHash: string = "";
-
-export async function getAgent() {
-    const proxy = await desktop.setting.net.getProxy();
-    const proxyHash = JSON.stringify(proxy);
-
-    if (cachedAgent && cachedProxyHash === proxyHash) {
-        return cachedAgent;
-    }
-
-    if (proxy && proxy.type !== "disabled" && proxy.host && proxy.port) {
-        let url: string;
-        let token: string | undefined;
-
-        if (proxy.type === "socks5") {
-            const auth =
-                proxy.requiresAuth && proxy.username && proxy.password
-                    ? `${encodeURIComponent(proxy.username)}:${encodeURIComponent(proxy.password)}@`
-                    : "";
-            url = `socks5://${auth}${proxy.host}:${proxy.port}`;
-        } else {
-            url = `http://${proxy.host}:${proxy.port}`;
-            if (proxy.requiresAuth && proxy.username && proxy.password) {
-                token = `Basic ${Buffer.from(`${proxy.username}:${proxy.password}`).toString("base64")}`;
-            }
-        }
-
-        cachedAgent = new ProxyAgent({
-            uri: url,
-            token,
-            factory(origin, options) {
-                return new Pool(origin, {
-                    ...options,
-                    allowH2: true,
-                });
-            },
-        });
-    } else {
-        cachedAgent = new Agent({
-            factory(origin, options) {
-                return new Pool(origin, {
-                    ...options,
-                    allowH2: true,
-                });
-            },
-        });
-    }
-
-    cachedProxyHash = proxyHash;
-    return cachedAgent;
-}
-
-export async function getHeaders(url: string) {
-    const token = await desktop.service.auth.getToken();
-    const prefixs = ["http://localhost", "https://api.nahida.live"];
-    const isNHD = prefixs.some((prefix) => url.startsWith(prefix));
-    return {
-        ...(token && isNHD && { Authorization: `Bearer ${token}` }),
-        Origin: "https://nahida.live",
-        "User-Agent": `Nahida Desktop/${appVersion}`,
-    };
-}
-
-export async function fetcher(url: string, options?: RequestInit) {
-    const prefixs = ["http://localhost", "https://api.nahida.live"];
-    const isNHD = prefixs.some((prefix) => url.startsWith(prefix));
-
-    const resp = await ky(url, {
-        ...options,
-        headers: {
-            ...options?.headers,
-            ...(await getHeaders(url)),
-        },
-        timeout: 100000,
-        retry: {
-            limit: 2,
-        },
-        // @ts-expect-error - dispatcher is not in the type definition, but it's passed through to fetch.
-        dispatcher: await getAgent(),
-        hooks: {
-            afterResponse: [
-                async (_req, _opt, resp) => {
-                    if (resp.status === 401 && isNHD) {
-                        await desktop.service.auth.getSession();
-                    }
-                },
-
-                (_request, _options, response) => {
-                    if (response.status === 524) {
-                        return new Response("cloudflare timeout. but it's ok", { status: 200 });
-                    } else {
-                        return response;
-                    }
-                },
-            ],
-
-            beforeError: [
-                async (error) => {
-                    const { response } = error;
-                    if (response && response.status === 524) {
-                        return error;
-                    }
-
-                    return error;
-                },
-            ],
-        },
-    });
-
-    return resp;
-}
