+++ src/main/services/auth.ts
@@ -1,6 +1,5 @@
 import type { NahidaDesktop } from "@main/index";
 import { setting } from "@main/internal/db/schema";
-import { fetcher, getAgent, getHeaders } from "@main/internal/fetcher";
 import { focus } from "@main/windows/utils";
 import { BACKEND_URL } from "@shared/const";
 import { SessionSchema } from "@shared/schemas/auth";
@@ -61,7 +60,7 @@ export class Auth {
         if (!token) return null;
 
         const url = `${BACKEND_URL}/api/auth/get-session`;
-        const resp = await fetcher(url);
+        const resp = await this.desktop.fetcher(url);
         const data = await resp.text();
         if (data === "null") {
             await this.startLogout();
@@ -83,9 +82,9 @@ export class Auth {
         const iWantToLoginResp = await ky(iWantToLoginUrl, {
             credentials: "include",
             throwHttpErrors: false,
-            headers: await getHeaders(iWantToLoginUrl),
+            headers: await this.desktop.getHeaders(iWantToLoginUrl),
             // @ts-expect-error - dispatcher is not in the type definition, but it's passed through to fetch.
-            dispatcher: await getAgent(),
+            dispatcher: await this.desktop.getAgent(),
         });
 
         if (!iWantToLoginResp.ok) {
@@ -103,9 +102,9 @@ export class Auth {
 
         const resp = await ky(data.stateResponse, {
             throwHttpErrors: false,
-            headers: await getHeaders(data.stateResponse),
+            headers: await this.desktop.getHeaders(data.stateResponse),
             // @ts-expect-error - dispatcher is not in the type definition, but it's passed through to fetch.
-            dispatcher: await getAgent(),
+            dispatcher: await this.desktop.getAgent(),
         });
 
         if (!resp.body) {
@@ -164,7 +163,7 @@ export class Auth {
         const token = await this.getToken();
         if (token) {
             const url = `${BACKEND_URL}/api/auth/sign-out`;
-            await fetcher(url, {
+            await this.desktop.fetcher(url, {
                 method: "POST",
             });
         }
